Add files via upload
We started by doing a quick quality control of the sequences provided in order to determine which parameters to use during trimming. 
We used the fastqc tool on each of the reads. 
The command used was: fastqc *.fastq.z -o qcresults/
In order to generate a general QC report, we did a multiqc analysis on the entire folder containing the generated fastqc  reports.
The command used was: multiqc qcresults/
This produced an html report attached here.